Added icons dir, rightArrow icons, updated tests & imports
