+++ src/helpers/icons.js
@@ -1,0 +1,107 @@
+import React from 'react';
+
+export const GettingStartedIcon = () => (
+  <svg
+    width="100%"
+    height="100%"
+    viewBox="0 0 28 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <rect width="28" height="28" rx="6" fill="#007AFC" />
+    <path
+      d="M21.9999 8.01145C21.9756 8.74628 21.8034 9.44521 21.5658 10.1295C21.1699 11.2693 20.6128 12.331 19.9533 13.3375C19.583 13.9025 19.1974 14.4591 18.7888 14.9964C18.4684 15.4177 18.1024 15.8041 17.7561 16.2057C17.6776 16.2968 17.5855 16.3793 17.5231 16.4804C17.4825 16.5462 17.4556 16.644 17.4712 16.7171C17.557 17.1204 17.6843 17.5168 17.7441 17.9234C17.7886 18.227 17.7075 18.5402 17.4832 18.7664C16.1173 20.1436 14.7448 21.5142 13.3677 22.8801C13.2935 22.9536 13.114 23.0274 13.0599 22.9902C12.9641 22.9243 12.894 22.7794 12.8711 22.6567C12.7693 22.1057 12.6869 21.5508 12.5998 20.9971C12.5317 20.5649 12.4486 20.1336 12.406 19.6987C12.3751 19.3838 12.4994 19.0942 12.6919 18.8458C12.8338 18.6629 12.8418 18.5788 12.6753 18.4119C11.6531 17.3888 10.6316 16.3647 9.60651 15.3445C9.42369 15.1627 9.33195 15.1663 9.1232 15.3189C8.83534 15.5294 8.51357 15.6428 8.13762 15.5657C7.76832 15.4898 7.3917 15.4499 7.01808 15.3931C6.43604 15.3046 5.85267 15.2232 5.27296 15.1198C5.17191 15.1018 5.05324 15.0094 5.00737 14.9186C4.97579 14.8564 5.05191 14.7357 5.08848 14.6446C5.10177 14.6114 5.14033 14.5884 5.16792 14.5608C6.46895 13.2597 7.77696 11.9656 9.06669 10.6532C9.44696 10.2662 9.87543 10.206 10.3644 10.3177C10.6672 10.3869 10.967 10.4707 11.2709 10.5338C11.4311 10.5674 11.5837 10.5175 11.7096 10.4108C11.8888 10.2589 12.06 10.0969 12.2415 9.94796C12.7205 9.55494 13.1905 9.14928 13.6878 8.78053C14.1747 8.41943 14.6863 8.08992 15.1966 7.76107C15.4927 7.57021 15.8049 7.40396 16.1143 7.23405C16.7608 6.87926 17.4293 6.57369 18.132 6.34426C18.6509 6.17468 19.1728 6.04102 19.7229 6.00876C20.1447 5.98382 20.5526 6.00643 20.9455 6.16504C21.4763 6.37917 21.7855 6.78516 21.9068 7.33413C21.956 7.55558 21.9995 7.73437 21.9999 8.01145Z"
+      fill="white"
+    />
+    <path
+      d="M5.8343 21.8561C5.94313 21.499 6.03433 21.1352 6.1638 20.7854C6.55586 19.7279 7.07541 18.7362 7.75703 17.8338C7.8885 17.6601 8.02696 17.4866 8.18938 17.3438C8.24862 17.2916 8.39673 17.3016 8.48592 17.3335C8.57712 17.366 8.65067 17.4561 8.72389 17.5288C9.27006 18.0714 9.82122 18.6092 10.3571 19.1614C10.7741 19.5913 10.7002 19.5425 10.3644 19.8744C9.78861 20.4435 9.11097 20.8435 8.3488 21.1232C7.74206 21.3458 7.15495 21.6219 6.55953 21.8747C6.42773 21.9306 6.29526 21.9857 6.16613 22.0468C5.91351 22.1661 5.77795 22.1163 5.8343 21.8561Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export const TutorialsIcon = () => (
+  <svg
+    width="100%"
+    height="100%"
+    viewBox="0 0 28 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <rect width="28" height="28" rx="6" fill="#2FC078" />
+    <path
+      d="M14 5C10.6863 5 8 7.8206 8 11.3C8 12.4142 8.28628 13.4753 8.80399 14.4221C8.81342 14.4401 10.4274 17.1032 10.5714 17.6H17.4286C17.57 17.1059 19.1789 14.4617 19.1694 14.4779C19.7034 13.5212 20 12.4322 20 11.3C20 7.8206 17.3137 5 14 5ZM14 8.6C14.4731 8.6 14.8571 9.0032 14.8571 9.5C14.8571 9.9968 14.4731 10.4 14 10.4C13.5269 10.4 13.1429 10.8032 13.1429 11.3C13.1429 11.7968 12.7589 12.2 12.2857 12.2C11.8126 12.2 11.4286 11.7968 11.4286 11.3C11.4286 9.8087 12.5797 8.6 14 8.6ZM10.5714 19.4C10.5714 21.3881 12.1066 23 14 23C15.8934 23 17.4286 21.3881 17.4286 19.4H10.5714Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export const HowToVideosIcon = () => (
+  <svg
+    width="100%"
+    height="100%"
+    viewBox="0 0 28 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <rect width="28" height="28" rx="6" fill="#9C67E7" />
+    <path
+      d="M18.8528 13.1848C19.4139 13.5835 19.4139 14.4165 18.8528 14.8152L11.8292 19.806C11.1671 20.2765 10.25 19.8031 10.25 18.9909L10.25 9.00914C10.25 8.19689 11.1671 7.72349 11.8292 8.19398L18.8528 13.1848Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export const TroubleshootingIcon = () => (
+  <svg
+    width="100%"
+    height="100%"
+    viewBox="0 0 28 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <rect width="28" height="28" rx="6" fill="#F13219" />
+    <path
+      d="M14.0123 11.5714C14.2602 11.3452 14.4851 11.1326 14.7188 10.9307C14.9004 10.7736 14.9591 10.5988 14.8968 10.3602C14.634 9.354 14.6429 8.36017 15.0718 7.39292C15.8264 5.69224 17.8708 4.66239 19.625 5.10114C19.7176 5.12417 19.8137 5.14957 19.8967 5.19504C20.1139 5.31373 20.1352 5.55584 19.943 5.75307C19.6642 6.03946 19.3777 6.31819 19.0947 6.59986C18.7013 6.99137 18.3074 7.38229 17.9159 7.77557C17.6578 8.03421 17.6554 8.24325 17.9123 8.5013C18.4658 9.05697 19.0211 9.61146 19.5782 10.1636C19.8303 10.4134 20.0569 10.414 20.3102 10.1636C20.9521 9.5282 21.5911 8.89104 22.2288 8.25152C22.3546 8.12574 22.4827 8.01708 22.6756 8.07673C22.863 8.1346 22.8951 8.29994 22.9259 8.46174C23.4076 10.9885 21.4962 13.6104 18.4818 13.2679C18.2107 13.2372 17.9408 13.1829 17.6738 13.1226C17.4395 13.0695 17.2562 13.1203 17.0996 13.3063C16.924 13.5153 16.7347 13.7126 16.5348 13.9334C14.0123 11.5714 14.8671 12.3721 14.0123 11.5714Z"
+      fill="white"
+    />
+    <path
+      d="M14.0123 11.5714L16.5348 13.9334C16.4274 14.0527 13.1338 17.6938 13.0312 17.8071C11.7065 19.271 10.3865 20.739 9.05521 22.197C7.94524 23.4129 6.07352 23.218 5.28924 21.8155C4.76777 20.8825 4.96295 19.7369 5.78164 18.9923C6.74449 18.1166 7.71505 17.2479 8.68146 16.3763C9.3293 15.7923 13.346 12.1731 14.0123 11.5714Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export const FaqIcon = () => (
+  <svg
+    width="28"
+    height="28"
+    viewBox="0 0 28 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <rect width="28" height="28" rx="6" fill="#1BC4FC" />
+    <path
+      d="M14 5C16.7665 5 19 6.81579 19 9.60526C19 11.3158 18.1371 12.3158 16.8934 13.4474C15.8782 14.3684 15.1168 14.9737 15.1168 15.7632C15.1168 15.9737 15.1421 16.1579 15.1675 16.3158H12.2741C12.1218 15.9737 11.9949 15.5 11.9949 15.0526C11.9949 13.7895 13.0609 12.8947 14 12.0789C15.0406 11.1579 15.4975 10.7632 15.4975 9.92105C15.4975 8.89474 14.736 8.26316 13.8985 8.26316C12.9594 8.26316 12.2741 8.94737 11.7919 10.2632L9 9C9.73604 6.31579 11.6142 5 14 5ZM13.8223 18.1053C15.1675 18.1053 16.1827 19.1842 16.1827 20.5526C16.1827 21.9474 15.1675 23 13.8223 23C12.4518 23 11.4365 21.9474 11.4365 20.5526C11.4365 19.1842 12.4518 18.1053 13.8223 18.1053Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export const GlossaryIcon = () => (
+  <svg
+    width="100%"
+    height="100%"
+    viewBox="0 0 28 28"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <rect width="28" height="28" rx="6" fill="#FC8200" />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M7 6.5C7 5.67157 7.67157 5 8.5 5H19.5C20.3284 5 21 5.67157 21 6.5V21.5C21 22.3284 20.3284 23 19.5 23H8.5C7.67157 23 7 22.3284 7 21.5V6.5ZM10 10C10 9.44772 10.4477 9 11 9H17C17.5523 9 18 9.44772 18 10C18 10.5523 17.5523 11 17 11H11C10.4477 11 10 10.5523 10 10ZM11 13C10.4477 13 10 13.4477 10 14C10 14.5523 10.4477 15 11 15H14C14.5523 15 15 14.5523 15 14C15 13.4477 14.5523 13 14 13H11Z"
+      fill="white"
+    />
+  </svg>
+);
